Refactor: factor out search.rs
